+++ appstack_plugin/ios/Sources/appstack_plugin/AppstackPlugin.swift
@@ -17,6 +17,10 @@ public class AppstackPlugin: NSObject, FlutterPlugin {
       handleSendEvent(call: call, result: result)
     case "enableAppleAdsAttribution":
       handleEnableAppleAdsAttribution(result: result)
+    case "getAppstackId":
+      handleGetAppstackId(result: result)
+    case "isSdkDisabled":
+      handleIsSdkDisabled(result: result)
     default:
       result(FlutterMethodNotImplemented)
     }
@@ -86,6 +90,16 @@ public class AppstackPlugin: NSObject, FlutterPlugin {
       AppstackASAAttribution.shared.enableAppleAdsAttribution()
     }
     result(true)
+  }
+  
+  private func handleGetAppstackId(result: @escaping FlutterResult) {
+    let appstackId = AppstackAttributionSdk.shared.getAppstackId()
+    result(appstackId)
+  }
+  
+  private func handleIsSdkDisabled(result: @escaping FlutterResult) {
+    let isSdkDisabled = AppstackAttributionSdk.shared.isSdkDisabled()
+    result(isSdkDisabled)
   }
   
   private func stringToEventType(_ string: String) -> AppstackSDK.EventType? {