+++ appstack_plugin/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to the Appstack Flutter Plugin will be documented in this fi
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.14] - 2025-10-31
+
+### Changed
+- Modified `configure()` method to return `Future<void>` instead of `Future<bool>`
+- Updated documentation and examples to reflect the void return type
+- Add a method `isSdkDisabled()` to see the value configured in the iOS or the Android SDK
+
 ## [0.0.13] - 2025-10-29
 
 ### Added